pyproject.toml
[tool.poetry]
name = "newcollegesite"
version = "0.1.0"
description = ""
authors = ["lionex-ui <eugenemescaline@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.38"
alembic = "^1.14.1"
fastapi = "^0.115.8"
flake8 = "^7.1.1"
isort = "^6.0.0"
asyncpg = "^0.30.0"
uvicorn = "^0.34.0"
setuptools = "^75.8.0"
flake8-bugbear = "^24.12.12"
flake8-comprehensions = "^3.16.0"
flake8-import-order = "^0.18.2"
black = "^25.1.0"
psycopg2-binary = "^2.9.10"
pytest = "^8.3.4"
httpx = "^0.28.1"
pytest-asyncio = "^0.25.3"
pydantic-settings = "^2.8.0"
sqlalchemy-utils = "^0.41.2"
pyjwt = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.isort]
profile = "black"
line_length = 120
